pyproject.toml
[project]
name = "flaskr"
version = "1.0.0"
description = "A basic web app for detecting text languages."

dependencies = [
    "flask",
    "langid",
    "iso639-lang",
    "python-dotenv",
    "httpx",
    "SQLAlchemy",
    "Flask-SQLAlchemy",
    "flask-security-too",
    "bcrypt"
]


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
python_requires = ">= 3.11"

[tool.setuptools]
packages = ["program"]

[tool.bandit]
exclude_dirs = ["tests"]


